Create a components dir with the components to be used   - App.js was moved to components   - Create Display, Button and ButtonPanel files
